# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sports Prop Bet Challenge Portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  open: "open",
  active: "active",
  completed: "completed",
} as const;

export interface Game {
  id: number;
  name: string;
  description?: string | null;
  status: GameStatus;
  adminCode: string;
  createdAt: string;
}

export type GameDetailStatus =
  (typeof GameDetailStatus)[keyof typeof GameDetailStatus];

export const GameDetailStatus = {
  open: "open",
  active: "active",
  completed: "completed",
} as const;

export type PropType = (typeof PropType)[keyof typeof PropType];

export const PropType = {
  yes_no: "yes_no",
  over_under: "over_under",
} as const;

export interface Prop {
  id: number;
  gameId: number;
  question: string;
  type: PropType;
  threshold?: number | null;
  result?: boolean | null;
  resolvedAt?: string | null;
  createdAt: string;
  order: number;
}

export interface GameDetail {
  id: number;
  name: string;
  description?: string | null;
  status: GameDetailStatus;
  adminCode: string;
  createdAt: string;
  props: Prop[];
  playerCount: number;
}

export interface CreateGameBody {
  name: string;
  description?: string | null;
}

export type UpdateGameBodyStatus =
  (typeof UpdateGameBodyStatus)[keyof typeof UpdateGameBodyStatus];

export const UpdateGameBodyStatus = {
  open: "open",
  active: "active",
  completed: "completed",
} as const;

export interface UpdateGameBody {
  status?: UpdateGameBodyStatus;
  name?: string;
  description?: string | null;
}

export type CreatePropBodyType =
  (typeof CreatePropBodyType)[keyof typeof CreatePropBodyType];

export const CreatePropBodyType = {
  yes_no: "yes_no",
  over_under: "over_under",
} as const;

export interface CreatePropBody {
  question: string;
  type: CreatePropBodyType;
  threshold?: number | null;
  order?: number;
}

export interface UpdatePropBody {
  result?: boolean | null;
  question?: string;
  order?: number;
}

export interface Player {
  id: number;
  gameId: number;
  name: string;
  joinedAt: string;
}

export interface JoinGameBody {
  name: string;
}

export interface Answer {
  id: number;
  propId: number;
  playerId: number;
  answer: boolean;
  createdAt: string;
}

export type SubmitAnswersBodyAnswersItem = {
  propId: number;
  answer: boolean;
};

export interface SubmitAnswersBody {
  playerId: number;
  answers: SubmitAnswersBodyAnswersItem[];
}

export interface LeaderboardEntry {
  playerId: number;
  playerName: string;
  score: number;
  correctAnswers: number;
  totalResolved: number;
  rank: number;
}

export interface Leaderboard {
  gameId: number;
  gameName: string;
  entries: LeaderboardEntry[];
  resolvedPropCount: number;
  totalPropCount: number;
}

export type GameSummaryStatus =
  (typeof GameSummaryStatus)[keyof typeof GameSummaryStatus];

export const GameSummaryStatus = {
  open: "open",
  active: "active",
  completed: "completed",
} as const;

export interface GameSummary {
  gameId: number;
  gameName: string;
  status: GameSummaryStatus;
  totalProps: number;
  resolvedProps: number;
  totalPlayers: number;
  topPlayer?: string | null;
  topScore?: number | null;
}